requirements.txt
certifi==2018.4.16
chardet==3.0.4
cloudinary==1.11.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-appconf==1.0.2
django-cloudinary-storage==0.2.3
django-image-cropping==1.1.0
djangorestframework==3.8.2
easy-thumbnails==2.5
funcsigs==1.0.2
gunicorn==19.8.1
heroku==0.1.4
idna==2.6
mock==2.0.0
pbr==4.0.2
Pillow==5.1.0
pkg-resources==0.0.0
psycopg2-binary==2.7.4
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.11.0
static3==0.7.0
urllib3==1.22
virtualenv==15.2.0
whitenoise==3.3.1
